+++ packages/ts-components/src/components/in-article-info-card/Card.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import {
+  CardContainer,
+  CardContent,
+  Label,
+  Headline,
+  HeadlineButtonContainer
+} from './styles';
+import { InfoCardData } from './InfoCard';
+
+export const Card: React.FC<{
+  headline: string;
+  label: string;
+  sectionColour: string;
+  data: InfoCardData;
+}> = ({ sectionColour, label, headline, children }) => (
+  <CardContainer>
+    <CardContent>
+      <Label sectionColour={sectionColour}>{label}</Label>
+      <HeadlineButtonContainer>
+        <Headline>{headline}</Headline>
+        {children}
+      </HeadlineButtonContainer>
+    </CardContent>
+  </CardContainer>
+);