+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.25.40](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.25.39...@times-components/ssr@2.25.40) (2021-08-02)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.25.39](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.25.38...@times-components/ssr@2.25.39) (2021-07-29)
 
 **Note:** Version bump only for package @times-components/ssr