+++ packages/ts-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.21](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.20...@times-components/ts-components@1.11.21) (2021-08-02)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
 ## [1.11.20](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.19...@times-components/ts-components@1.11.20) (2021-07-29)